fixed ReturnHandler and its tests
